+++ plugins/files/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @matatbread/matbot-files-node
 
+## 0.2.8
+
+### Patch Changes
+
+- @matatbread/matbot-core@0.2.8
+
 ## 0.2.7
 
 ### Patch Changes