+++ plugins/frontend/web/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @matatbread/matbot-frontend-web
 
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @matatbread/matbot-plugin-api@0.2.8
+  - @matatbread/matbot-core@0.2.8
+
 ## 0.2.7
 
 ### Patch Changes