+++ plugins/mcp/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @matatbread/matbot-tool-mcp
 
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @matatbread/matbot-plugin-api@0.2.8
+  - @matatbread/matbot-mcp-http@0.2.8
+
 ## 0.2.7
 
 ### Patch Changes